Update to OsmAnd 3.2
Fixes #1344

It is possible that the build with fail with "aidl not found".  This is
causes by the build.gradle in OsmAnd-core/wrappers/android specifying an
out-of-date build tools version.  Gradle automatically picks a
"compatible" version instead.  On my system this caused a failure until
i uninstalled build tools 27.0.3.  I have tested with 28.0.2, but did
not change the build.gradle as perhaps a better choice is made
automatically on the F-Droid build server.