Restore proper license tags
The CC licenses are correct for those, since we have to list to the license
for APK, not pieces of the source repo. It is the license of the
combination of all the licenses of all the pieces put together. The most
restrictive license wins, in this regard.

https://gitlab.com/fdroid/fdroiddata/merge_requests/5623#note_226734466

This reverts 3/4 of changes in commit 07b3b8c33b0dbb22c059079c075bf254e16ef280.