Acrylic Paint: update repository URL to Phabricator
Since a couple of years the main repository of Acrylic Paint is
the one at gitpull.it and not on GitHub.

This can be verified visiting the legacy official repository:
	https://github.com/valerio-bozzolan/AcrylicPaint

And looking at the URL at the right sidebar quoting "gitpull.it".

The issue tracker will be migrated also soon but not now because
people still use GitHub.

Anyway, the current maintainer do not trust GitHub, so that's why
the source code is hosted elsewhere on the personal infrastructure
of the current maintainer.