Update AnySoftKeyboard language packs to 4.0.1396 (2908)
A couple language packs use a different version in order to match APK at
Play Store.

This PR was amended several times. This commit was achieved by squashing
and rebasing all previous ones, and is the first to pass lint and build.

Hungarian is renamed to `hungarian_oss` and Dutch to `dutch_oss`, using
`afrikaans_oss` as a template. Metadata files have been fixed and licence
for Hungarian language pack (which is technically triple/quadruply licenced)
has been retained.